+++ src/lib.rs
@@ -5,6 +5,7 @@ extern crate diesel_migrations;
 
 pub mod app;
 mod database;
+mod kafka_consumer;
 mod server;
 
 use anyhow::Error;