+++ CHANGELOG.md
@@ -8,6 +8,15 @@ project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 - Added `event` and `kafka_metadata` tables.
+- Added a new endpoint `GET /api/event/search` to fetch raw_events from event
+  table.
+- Added periodic tasks to fetch kafka metadata and raw events that work with the
+  environment below. `raw_events` corresponding to `event_ids` in `cluster`
+  and `outlier` tables are stored in `event` table by these periodic tasks.
+    - `MAX_OFFSET_COUNT`: Maximum number of offsets for each kafka topic to
+      fetch per task. Default value is 1000.
+    - `TASK_TIME_INTERVAL`: Periodic task interval (in seconds). Default
+      value is 900 (15 minutes).
 
 ## [0.6.3] - 2019-12-30
 
@@ -17,7 +26,8 @@ project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 
-- Enum values are stored as `String` type of raw data instead of their mapped unsigned integers.
+- Enum values are stored as `String` type of raw data instead of their mapped
+  unsigned integers.
 
 ### Deprecated